DF-1660 DF-1659 moved backlists components
