+++ alv-portal-ui/src/app/competence-catalog/competence-elements/competence-elements.module.ts
@@ -7,7 +7,6 @@ import { CompetenceElementsOverviewComponent } from './competence-elements-overv
 import { InfiniteScrollModule } from 'ngx-infinite-scroll';
 import { SharedCompetenceCatalogModule } from '../shared/shared-competence-catalog.module';
 import { CompetenceElementsFilterModalComponent } from './competence-elements-filter-modal/competence-elements-filter-modal.component';
-import { CompetenceElementBacklinksComponent } from './competence-elements-overview/competence-element-backlinks/competence-element-backlinks.component';
 
 @NgModule({
   imports: [
@@ -20,11 +19,9 @@ import { CompetenceElementBacklinksComponent } from './competence-elements-overv
   declarations: [
     CompetenceElementsOverviewComponent,
     CompetenceElementsFilterModalComponent,
-    CompetenceElementBacklinksComponent
   ],
   entryComponents: [
     CompetenceElementsFilterModalComponent,
-    CompetenceElementBacklinksComponent
   ],
   providers: [
     ModalService
