+++ alv-portal-ui/src/app/competence-catalog/competence-sets/competence-sets.module.ts
@@ -8,7 +8,6 @@ import { SharedCompetenceCatalogModule } from '../shared/shared-competence-catal
 import { InfiniteScrollModule } from 'ngx-infinite-scroll';
 import { CompetenceSetDetailComponent } from './competence-set-detail/competence-set-detail.component';
 import { CompetenceElementSearchModalComponent } from './competence-element-search-modal/competence-element-search-modal.component';
-import { CompetenceSetBacklinksComponent } from './competence-sets-overview/competence-set-backlinks/competence-set-backlinks.component';
 
 @NgModule({
   imports: [
@@ -22,11 +21,9 @@ import { CompetenceSetBacklinksComponent } from './competence-sets-overview/comp
     CompetenceSetsOverviewComponent,
     CompetenceSetDetailComponent,
     CompetenceElementSearchModalComponent,
-    CompetenceSetBacklinksComponent
   ],
   entryComponents: [
     CompetenceElementSearchModalComponent,
-    CompetenceSetBacklinksComponent
   ],
   providers: [
     ModalService
