+++ alv-portal-ui/src/app/competence-catalog/shared/backlinks/competence-element-backlinks/competence-element-backlinks.component.ts
