+++ alv-portal-ui/src/app/competence-catalog/shared/backlinks/competence-set-backlinks/competence-set-backlinks.component.ts
